Add files via upload
Rock-Paper-Scissors-Lizard-Spock is a tournament, so each winner knows his opponents in each round. During the tournament, players are not allowed to change their signs. The number of tournament games is equal to the number of players/2. In case both players have the same sign a coin toss should decide the winner.  Also, at the end there is a matplotlib barchart that shows the number of each sign chosen during the game.